[optimize]優化firestore讀取&寫入用量控制
