+++ app/(main)/chat/page.tsx
@@ -3,19 +3,23 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 
 import React, { useEffect, useState } from 'react';
-import { useAppSelector } from '@/store/hooks';
+import { useAppDispatch, useAppSelector } from '@/store/hooks';
 import ChatRoom from '@/components/ChatRoom';
 import ChatList from '@/components/ChatList';
 import { RootState } from '@/store';
-import { getMessages, updateReadStatus } from '@/lib/chat';
-import { messageDataType } from '@/types/chatType';
+import { getChatList, getMessages, updateReadStatus } from '@/lib/chat';
+import { chatListInfoType, messageDataType } from '@/types/chatType';
 import { useMessage } from '@/hooks/useMessage';
+import { setChatList } from '@/store/chatSlice';
+import { setUnReadMessageCount } from '@/store/sysSlice';
 
 function Chat() {
+  const dispatch = useAppDispatch();
   const activeChatRoomId = useAppSelector((state) => state.chat.activeChatRoom?.chatRoomId);
   const [currentRoomId, setCurrentRoomId] = useState<string | undefined>(activeChatRoomId);
   const [messageList, setMessageList] = useState<messageDataType[]>([]);
   const uid = useAppSelector((state: RootState) => state.user.userData?.uid);
+  const chatList = useAppSelector((state: RootState) => state.chat.chatList);
 
   /** 取得聊天室訊息資料 */
   const handleGetMessage = async (roomId: string, activeRoomId: string) => {
@@ -28,10 +32,25 @@ function Chat() {
     }
   };
 
+  /** 取得聊天室列表資料 */
+  const handleGetChatList = async () => {
+    if (!uid) return;
+    const result = await getChatList(uid!);
+    if (result.code === "success") {
+      dispatch(setChatList(result.chatList as unknown as chatListInfoType[]));
+      const count = result.chatList?.reduce((acc, item) => acc + item.unreadCount, 0) || 0;
+      dispatch(setUnReadMessageCount(count));
+    }
+  };
+
   /** 更新讀取狀態 */
-  const handleUpdateReadStatus = async () => {
-    if (!currentRoomId || !uid) return;
-    await updateReadStatus(currentRoomId, uid);
+  const handleUpdateReadStatus = async (isSendMessage: boolean = false) => {
+    if (!currentRoomId || !uid || !chatList) return;
+    const currentRoomUnRead =  chatList.find((item) => item.chatRoomId === currentRoomId)?.unreadCount;
+    if ((currentRoomUnRead && currentRoomUnRead > 0) || isSendMessage) {
+      await updateReadStatus(currentRoomId, uid);
+      handleGetChatList();
+    }
   };
 
   useEffect(() => {
@@ -49,7 +68,9 @@ function Chat() {
   return (
     <div className="flex w-full h-full md:pt-5 md:px-5">
       <aside className="hidden md:block w-full md:max-w-60 lg:max-w-80 bg-[var(--card-bg-color)] md:rounded-tl-lg p-5">
-        <ChatList />
+        <ChatList
+          handleGetChatList={handleGetChatList}
+        />
       </aside>
       <section className="hidden md:block w-full h-full border-l border-[var(--divider-color)] bg-[var(--card-bg-color)] md:rounded-tr-lg">
         <ChatRoom
@@ -62,7 +83,9 @@ function Chat() {
       {/* 手機版 */}
       {(!activeChatRoomId || activeChatRoomId === "") ? (
         <div className="md:hidden w-full h-full bg-white dark:bg-[var(--background)] border-b border-[var(--divider-color)] pt-2 px-5">
-          <ChatList />
+          <ChatList
+            handleGetChatList={handleGetChatList}
+          />
         </div>
       ) : (
         <div className="md:hidden w-full h-full bg-white dark:bg-[var(--background)] border-b border-[var(--divider-color)]">
